crawl/requirements.txt
pillow
flask
flask-compress
requests
numpy
